Misc, Code clean-up
This commit is the start of the DAF addon loading, which at the moment is bare-bones, so only checking a few folders and files, and does nothing with it, only displaying it in the BetterConsole menu.

Includes renaming of files, such as rename the confusing "Scale" to a more understandable "Data Manipulator".

General code clean-up has started, including removing obsolete or unusable code, and commenting some for future adjustments.

Aside, adds BetterConsole for the menu.